move xla_path() to jax.py
